# backend/app/schemas/game.py
"""Pydantic DTOs for the game-logs API."""

from datetime import date
from typing import Any

from pydantic import BaseModel, ConfigDict


class GameLogOut(BaseModel):
    model_config = ConfigDict(from_attributes=True)

    game_id: int
    game_date: date
    opponent_id: int | None
    opponent_name: str | None = None
    opponent_abbrev: str | None = None
    is_home: bool | None
    level: str | None = None
    group_name: str
    stats: dict[str, Any]