# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Korean MLB Tracker API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/players": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "List Players",
        "description": "List tracked players, optionally filtered by ``level`` (MLB/AAA/AA/A+/A/R).",
        "operationId": "list_players_api_v1_players_get",
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlayerOut"
                  },
                  "title": "Response List Players Api V1 Players Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{player_id}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Get Player",
        "operationId": "get_player_api_v1_players__player_id__get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/players/{player_id}/games": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "List Player Games",
        "operationId": "list_player_games_api_v1_players__player_id__games_get",
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameLogOut"
                  },
                  "title": "Response List Player Games Api V1 Players  Player Id  Games Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboard/today": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Dashboard Today",
        "operationId": "dashboard_today_api_v1_dashboard_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardOut"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DashboardGameOut": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "full_name_ko": {
            "type": "string",
            "title": "Full Name Ko"
          },
          "full_name_en": {
            "type": "string",
            "title": "Full Name En"
          },
          "current_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Level"
          },
          "player_type": {
            "type": "string",
            "title": "Player Type"
          },
          "opponent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Id"
          },
          "opponent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Name"
          },
          "opponent_abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Abbrev"
          },
          "is_home": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Home"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "full_name_ko",
          "full_name_en",
          "current_level",
          "player_type",
          "opponent_id",
          "is_home",
          "group_name",
          "stats"
        ],
        "title": "DashboardGameOut",
        "description": "One tracked player's line in the latest game day's feed (player + game)."
      },
      "DashboardOut": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "games": {
            "items": {
              "$ref": "#/components/schemas/DashboardGameOut"
            },
            "type": "array",
            "title": "Games"
          }
        },
        "type": "object",
        "required": [
          "date",
          "games"
        ],
        "title": "DashboardOut",
        "description": "Latest game day plus every tracked player who appeared that day.\n\n``date`` is null when there are no game logs at all."
      },
      "GameLogOut": {
        "properties": {
          "game_id": {
            "type": "integer",
            "title": "Game Id"
          },
          "game_date": {
            "type": "string",
            "format": "date",
            "title": "Game Date"
          },
          "opponent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Id"
          },
          "opponent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Name"
          },
          "opponent_abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Opponent Abbrev"
          },
          "is_home": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Home"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "game_date",
          "opponent_id",
          "is_home",
          "group_name",
          "stats"
        ],
        "title": "GameLogOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PlayerDetailOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name_en": {
            "type": "string",
            "title": "Full Name En"
          },
          "full_name_ko": {
            "type": "string",
            "title": "Full Name Ko"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "player_type": {
            "type": "string",
            "title": "Player Type"
          },
          "current_team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Team Id"
          },
          "current_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Level"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "bats": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bats"
          },
          "throws": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throws"
          },
          "birth_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "season_stats": {
            "items": {
              "$ref": "#/components/schemas/SeasonStatsOut"
            },
            "type": "array",
            "title": "Season Stats"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name_en",
          "full_name_ko",
          "position",
          "player_type",
          "current_team_id",
          "current_level",
          "is_active",
          "bats",
          "throws",
          "birth_date",
          "season_stats"
        ],
        "title": "PlayerDetailOut",
        "description": "Player detail: summary plus bio and season totals."
      },
      "PlayerOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name_en": {
            "type": "string",
            "title": "Full Name En"
          },
          "full_name_ko": {
            "type": "string",
            "title": "Full Name Ko"
          },
          "position": {
            "type": "string",
            "title": "Position"
          },
          "player_type": {
            "type": "string",
            "title": "Player Type"
          },
          "current_team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Team Id"
          },
          "current_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Level"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name_en",
          "full_name_ko",
          "position",
          "player_type",
          "current_team_id",
          "current_level",
          "is_active"
        ],
        "title": "PlayerOut",
        "description": "Player summary for the list endpoint / cards."
      },
      "SeasonStatsOut": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          }
        },
        "type": "object",
        "required": [
          "season",
          "group_name",
          "level",
          "stats"
        ],
        "title": "SeasonStatsOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}